Moved footerContent to im-components
