+++ packages/im-components/src/organisms/footerContent/index.js
